+++ src/parser/enrichers/class/warlock/EldritchInvocations.mjs
@@ -1,0 +1,10 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../../data/DDBEnricherData.mjs";
+
+export default class EldritchInvocations extends DDBEnricherData {
+
+  get type() {
+    return "none";
+  }
+
+}