+++ src/parser/enrichers/mixins/DDBActivityFactoryMixin.mjs
@@ -429,8 +429,8 @@ export default class DDBActivityFactoryMixin {
       const activity = this.data.system.activities[activityId];
       if (!activity.effects || activity.effects.length !== 0) continue;
       if (foundry.utils.getProperty(activity, "flags.ddbimporter.noeffect")) continue;
-      const ignoreTransfer = foundry.utils.getProperty(activity, "flags.ddbimporter.ignoreTransfer") ?? false;
       for (const effect of this.data.effects) {
+        const ignoreTransfer = foundry.utils.getProperty(effect, "flags.ddbimporter.ignoreTransfer") ?? false;
         if (effect.transfer && !ignoreTransfer) continue;
         if (foundry.utils.getProperty(effect, "flags.ddbimporter.noeffect")) continue;
         const activityNamesRequired = foundry.utils.hasProperty(effect, "flags.ddbimporter.activitiesMatch")